+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Fixed
 - Handle `--server http://localhost:1234`
+- Tag and branch name validity checks match `zeitgitterd`'s
 
 ## Changed
 - `ZEITGITTER_FAKE_TIME` now used for testing