+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `--quiet` option
 
 ## Fixed
+- Handle `--server http://localhost:1234`
+- Tag and branch name validity checks match `zeitgitterd`'s
 
 ## Changed
 - `ZEITGITTER_FAKE_TIME` now used for testing