+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
-# 0.9.6+ - [Unreleased]
+# 0.9.7 - 2019-09-01
 ## Added
 - `--quiet` option